requirements.txt
elasticsearch>=5.0.0,<6.0.0
kafka-python
python-dateutil
requests
